automake silent rules: handle tsc
We were showing the full cmdline in case the --enable-silent-rules
configure option is used to just display "CXX      unithttplib-test.o"
style output.

The generated Makefile has a AM_DEFAULT_VERBOSITY variable based on the
configure switch, also V=1 on the cmdline can be used to enable
verbosity in silent builds.

Add a similar QUIET_TSC variable for the TypeScript compiler and used it
in all places where we invoke tsc.

This way the silent output emits less noise and we still have a way to
see the actual cmdline without touching the Makefile.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I692edacf74f4c5d86eb098e58c6493b16a410166