Use better checks for determining if a comment is focused or not. Some browsers don't support events like focusin focusout, also browsers may have different behaviour about parent of a focused element.
The issue: commentHasFocus variable is not updated correctly. This prevents cut/paste operations. Because when a comment is focused, cut/paste from document is prevented.
Fix: It's better to check if a modify / reply field is focused via document.activeElement.

This changes the code introduced in below commit:
* https://github.com/CollaboraOnline/online/pull/7634/files

So the fix will be re-checked.

Signed-off-by: Gökay Şatır <gokaysatir@collabora.com>
Change-Id: Ic6dd5d784af1c8cf5d98cd44dfbb8ac07ae045a3