cool#10630 doc electronic sign: send selected country to the server
So that in case e.g. Germany is selected, then the popup shows the
German flag. Having the flag there is not too important, but the popup
defaults to showing an Estonian flag otherwise, which is confusing.

The list of providers is not yet reduced based on the selected country.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ie39298d45909a3eec06ff571b422c6a0a7b102cf