Issue: There is a JS error when we scroll while mouse pointer is above a comment.
Fix: Simplify event propagation from comment container to CanvasSectionContainer.

Now we call sectionContainer's event from comment container.

Why we need it: Comment container is an HTML object and we want to selectively propagate events from it to canvas.
Note: The coordinate of action is probably wrong but ScrollSection is not using the mouse coordinates, so it's fine.

Signed-off-by: Gökay Şatır <gokaysatir@collabora.com>
Change-Id: If2819146c9d055b419944156cef12798b8974de8