cool#10630 doc electronic sign: set defaults without an index
The esign dialog is currently not sorting the country & provider list,
also providers are listed even if they won't work with a given country.

All this filtering & sorting is difficult to do if we have a full list
of those and we work with offsets into the full list.

A better way is to work with provider IDs and country codes, which are
stable across filtering & sorting.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I2aef2613b7c2bf4c5f426502de25a25c54dc9225