ci(iOS): bump build version
Due to an expired xcode login session on our CI, I noticed that the version
of xcode we were running was incompatible with the version of MacOS we were
running.

To rectify this, I updated xcode. Unfortunately, the latest versions of xcode
don't seem to have support for iOS 14.5 builds. Updating the deploy target to
the earliest supported version (16.3) fixes the issue.

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I0a6d4de2afda366019749dd1f6e0b51156c746be