slideshow: followup: exit  from notes view before starting
This is a follow up since previous patch has a problem when presenter console is
started in Firefox when Impress is in notes view mode.
Oddly, it seems that presentation information are requested but the reply with
the info never reaches CanvasTileLayer._onMessage.
Anyway I can see that the reply is sent by the core.
Since I'm not able to understand the root cause, I implemented a simple
workaround by delay the presenter console start by 500ms.
In order to be safer I decided to apply the same delay also when we present
without the console.

Adapt patch to LibreOffice Online.

Signed-off-by: Marco Cecchetti <marco.cecchetti@collabora.com>
Change-Id: I941419b3d33095608973af18aab9df2b767a02f4