Fix width of Thesaurus left arrow button
SVG images may not have an intrinsic width and height defined.
This can cause max-content to behave unexpectedly since it
relies on the size of the content, and an SVG without dimensions
may not contribute to the content's width.

We set width explicitely in this case.

Adapt patch to LibreOffice Online.

Signed-off-by: Andras Timar <andras.timar@collabora.com>
Change-Id: I5ebed738bd4f3a12a96596ddac07341097e43bc3