calc: improve v-scroll calculation with split panes
With 0f461ed3129250f6ef5fe003e6d543b3612f8da7
calc: scroll to the cell so it is visible

Vertical scrolling in case of split panes was designed to put
cell at the center of free pane. That doesn't take into account
cells higher than free pane. Let's do it in 2 steps, first
scroll to the top of a cell, then make sure bottom is visible.
That is similar behavior like in the desktop app.
Also don't use as a margin the height of a current cell but rather
as const value.

Signed-off-by: Szymon Kłos <szymon.klos@collabora.com>
Change-Id: Ic8ad463545dff372f953f454a9b19e6718672318