cool#10630 doc electronic sign: add a snackbar
In order to combine visual signing with electronic signing, you had to
use two menu items in the right order, which is possibly hard to
remember and is definitely not intuitive.

Also, it makes little sense to do non-visual signing when visual signing
works.

Fix this by merging to two menu items: there is now just a single insert
-> signature menu item that inserts the signature line shape. Also a
snackbar is shown that has a button to finish the signing (which does
the actual electronic signing).

This is quite similar to how this works on the desktop with infobars.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ic0f2490770540e1bcf0cf982d0cab6fdd5821e2b