cool#10630 doc electronic sign: remember the provider in localstorage
Insert -> Signature, the popup asks for your country & the signature
provider to use. This is typically the same for a user all the time, but
we keep asking for it & the user has to state it again & again.

The WOPI user info doesn't trivially provide the country, so just
showing a default on first use is reasonable.

Fix the problem by keeping the default unchanged, but store the user
choice in local storage, so later sign sessions will just show the popup
and the user can typically just accept the last used values.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Iba37f77e382ebd7421d9e7e525ba919397fb1596