cool#10630 doc electronic sign: show busy overlay while the popup is open
Insert -> Signature, then a popup browser window starts to handle
signing, but in case the user switches back to COOL, it's still possible
to e.g. move the signature widget.

The popup is used so we don't have to leave the edit session & return to
it, but otherwise it's assumed that the user doesn't interact with the
document while the sign popup is open.

Fix this by showing a "busy popup" inside the COOL UI till the popup is
open.

Given that hiding / showing this "busy popup" is an async event, make
sure that the UNO commands we dispatch during sign are executed even if
the hidebusy event is not yet processed.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I0ed945f56a34c145c21037c7d6c5077187133dbc