cool#10630 doc electronic sign: avoid unwanted postmessage during sign
Insert -> Signature, and once the sign popup shows up, the dev tools
console has this output in a 'make run' session:

==== debug.html receiveMessage: {"MessageId":"UI_Hyperlink","SendTime":1737387748712,"Values":{"Url":"https://test.eideasy.com/single-method-signature?client_id=2IaeiZXbcKzlP1KvjZH9ghty2IJKM8Lg&doc_id=NflXWVHfNoUucauIMlhUCDyxdEfpbQEhn1pe5dOS&method=smart-id-signature&country=EE","Name":"_blank","Features":"popup, left=480, top=270, width=960, height=540"}}

But this is not a useful notification for the integrator, since the URL
is not useful in itself, it can only work if the popup can do a
postmessage back to the COOL frame, which won't work whtn the URL is
opened in a separate way.

Just don't notify integrators about the popup's URL for now to avoid the
noise.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Ic72d9de902f1f2d570e09331372aa5bcaa8f6510