cool#10630 doc electronic sign: fix outdated tile for shape move on 2nd page
Open a PDF with 2 pages, go to the second page, insert -> signature
line, drag the graphic selection, the shape doesn't show up at the new
location.

Seems this is really similar to the problem when inserting a signature
line doesn't invalidate properly.

There is still no precise tracking of invalidation for read-only &
multi-page PDF files, so just extract Control.Menubar.js code to a new
requestNewFiledBasedViewTiles() and use that in
ShapeHandlesSection.sendTransformCommand().

With this, visual esign for non-first pages is about as usable as for
first pages, finally.

Adapt patch to LibreOffice Online.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: Idbf0cfe32030471eb0472ff9a00ccc0304cc4cf3