Notebookbar: Shorten "Find and replace" Translatable string
"Find and replace" button in the notebookbar can be:

- Too long: Requiring additional scrolling in narrower windows can
become particularly frustrating when dealing with certain long-form
languages."
- Too hard for first eye scanning: Since the button is placed within
the Search group ("Search" btn at the top and "Find and replace" btn
at the bottom) it can be difficult to spot "replace" when user tries
to find it but at the first sight sees only "Search" and then "Find".
- A bit unexpected, specially when the user is used to see "Replace"
in other workflows and Office suites.

Signed-off-by: Pedro Pinto Silva <pedro.silva@collabora.com>
Change-Id: I3e875191490610df5fbf717e259cf7c7bb85d61a