js: fix null exception in _onRefresh due to missing event removal
Fixes an issue where _onRefresh in Control.Menubar.js throws a null
exception because the updatepermission event listener was not removed
when removeClassicUI was called.

Issue:
1. The client initializes basicUI, but at this stage, the browsersetting
message has not been received yet.
2. Since useBrowserSetting is false, it may fallback to localStorage,
which could have "compactMode=true", leading to the initialization of
the menubar.
3. Later, when the browsersetting JSON is received, it determines the
final UI mode. If notebookbar mode is preferred, the menubar and top
toolbar are removed.
4. However, while removing events, the updatepermission event was not
detached. As a result, when updatepermission fired, _onRefresh was still
called on a now-null _map, causing an exception.

Fix:
- Ensure that the updatepermission event listener is properly removed
when removeClassicModeUI is executed.

Signed-off-by: Rashesh <rashesh.padia@collabora.com>
Change-Id: Ic5dc86c6b5350670e586966226ba722135a55b0d