Revert "wsd: shutdown and close sockets early"
This reverts commit 3e7dd1f7feaee7335d6de7e7701860f7b406010f.

Revert "wsd: simplify shutting down in checkRemoval"

This reverts commit 7b6a1186cf6cae9ab66e7976fcc0881ada566d41.

Revert "wsd: invalidate socket FD by negating"

This reverts commit 9c323376bf86d1b9abf8a3644eaf4457c2edc586.

Revert "wsd: replace _closed member with invalidating the FD"

This reverts commit af747c19fc96ae2024a5660034a23b7936620b25.

So we restore the traditional separation of ::shutdown on a fd
and the setting of a "close" stated in our Socket class.

Motivation is that bgsv depends on this separation, and fixing that
alone may not be sufficient for all other currently situations.

Which restores to a point where there may be occasional cases of fd
getting unexpectedly closed.

Signed-off-by: Caolán McNamara <caolan.mcnamara@collabora.com>
Change-Id: Icf91adcad45d8f58372e9beb8b8f5ca5b328ac18