make run: don't build all the C++ tests
Touching a C++ header or source file, make run, reload in browser is a
typical incremental developer cycle, which should be fast.

Integration tests were built in this case, but the rebuild of these is
not needed to run coolwsd.

Fix the problem by adding our injected .so integration tests to the
"check" automake category instead of the "noinst" one, so when 'make
run' depends on 'make all', we don't rebuild these, leading to a faster
development cycle.

Now the opposite problem would be that there is no easy way to just
build all C++ code: 'make' doesn't build tests and 'make check' builds &
runs tests. So introduce a new environment variable for this special
case: `make -C test check SUPPRESS_TESTS=y` can be now used to build
such test code without running it. The behavior for the test/unittest
unit test is unchanged, that is handled via an explicit rule in the
all-local target of test/Makefile.am.

Signed-off-by: Miklos Vajna <vmiklos@collabora.com>
Change-Id: I336549eeb09bd093e15fcb7943bbd4be96d2f3d2