chore: install @types/hammerjs
Previously we didn't have typescript types for hammerjs, so in our
typescript files it was always given as "any". I was writing some more
hammerjs code, noticed this and saw that there was already a
definitelytyped (@types/) package on npm, so there's little reason not
to install it!

The types provided by this package are a little weird - things like
having to use HammerManager rather than Hammer when typing - but they
are a great deal better than having no types at all...

Adapt patch to LibreOffice Online.

Signed-off-by: Skyler Grey <skyler.grey@collabora.com>
Change-Id: I7ac9bd6f390695c5b094aa89f948cf1dfe16df46